# [RAR] Don't do I/O on SDK-provided references