--- a/src/Tasks/AssemblyDependency/ReferenceTable.cs
+++ b/src/Tasks/AssemblyDependency/ReferenceTable.cs
@@ -1254,7 +1254,7 @@ namespace Microsoft.Build.Tasks
             Reference reference)
         {
             bool isImmutableFrameworkReference = false;
-            if (ChangeWaves.AreFeaturesEnabled(ChangeWaves.Wave17_7))
+            if (ChangeWaves.AreFeaturesEnabled(ChangeWaves.Wave17_8))
             {
                 // For a path to be an immutable reference, it must be externally resolved and has a FrameworkReferenceName defined.
                 if (assemblyName == null && !string.IsNullOrEmpty(rawFileNameCandidate) && reference.IsPrimary && reference.ExternallyResolved)