… documentation/wiki/ChangeWaves.md
… src/Tasks/AssemblyDependency/Resolver.cs
… src/Tasks/InstalledSDKResolver.cs