@ladipro 2023-05-04
I have addressed feedback and run another experimental insertion. It flagged several major improvements in, quite surprisingly, the add-new-project scenario (-30% reference set, for example).

@rokonec noticed that we still open SDK assemblies to read their metadata version ("runtime version"), which doesn't look necessary. I made `GetAssemblyRuntimeVersion` return the hardcoded `v4.0.30319` to remove this I/O, with the following justification:
- None of the assemblies currently distributed with the SDK/workloads appear to have anything else than this canonical version.
- Daniel is not aware of an issue with this assumption.
- The metadata parsing code has a bug that prevents it from detecting WinRT/winmd files, which would have been the only thing potentially broken by the change (#8731). I.e. _it doesn't work anyways_.

@Forgind @rainersigwald, please review. Merging this PR will unblock further RAR perf work.